Fix JNI types according to FreeType documentation
Advance should be `long`, not `int`. This can affect certain devices, making the advance width/height `0`.

+++ mathdisplaylib/src/main/java/com/pvporbit/freetype/GlyphSlot.java
@@ -53,18 +53,18 @@ public class GlyphSlot extends Pointer {
 
     public static class Advance {
 
-        private final int x, y;
+        private final long x, y;
 
-        public Advance(int x, int y) {
+        public Advance(long x, long y) {
             this.x = x;
             this.y = y;
         }
 
-        public int getX() {
+        public long getX() {
             return x;
         }
 
-        public int getY() {
+        public long getY() {
             return y;
         }
